Seperate Widget showing grid of bytes from rest of the hexdump panel.
